Fix player motion (it was supposed to be handled outside event handling)
please grade homework6 windows

+++ src/gamedata/Direction.h
@@ -4,8 +4,8 @@
 
 class Direction {
 public:
-	static inline const glm::vec2 UP = glm::ivec2(0, -1);
-	static inline const glm::vec2 DOWN = glm::ivec2(0, 1);
-	static inline const glm::vec2 LEFT = glm::ivec2(-1, 0);
-	static inline const glm::vec2 RIGHT = glm::ivec2(1, 0);
+	static inline const glm::vec2 UP = glm::vec2(0, -1);
+	static inline const glm::vec2 DOWN = glm::vec2(0, 1);
+	static inline const glm::vec2 LEFT = glm::vec2(-1, 0);
+	static inline const glm::vec2 RIGHT = glm::vec2(1, 0);
 };
